+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # KeyboardKit change log
 
 - Next
+    - Xcode 14 is now required.
+    - Adds support for iOS 16.
+    - Disables `KeyboardTextView`’s find commands when the text view’s built-in `UIFindInteraction` is enabled. 
     - Fixes adjustments to the date of `KeyboardDatePicker` within the same month being animated on iOS 16, which made it feel less responsive.
 - 6.0.0
     - Xcode 13.1 is now required.